// app/[locale]/(app)/watch/services/types.ts
export interface Video {
  id: string;
  title: string;
  channelTitle: string;
  viewCount: string;
  publishedAt: string;
  duration: string;
  thumbnailUrl: string;
}

export interface VideoDetails {
  id: string;
  title: string;
  channelTitle: string;
  viewCount: string;
  publishDate: string;
  description: string;
  subscriberCount: string;
  likeCount: string;
  dislikeCount: string;
}